Move useCustomFields from smartSearch feature to profile feature
