+++ src/features/profile/store.ts
@@ -1,12 +1,19 @@
-import { ZetkinPerson } from 'utils/types/zetkin';
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RemoteItem, remoteItem } from 'utils/storeUtils';
+import {
+  RemoteItem,
+  remoteItem,
+  RemoteList,
+  remoteList,
+} from 'utils/storeUtils';
+import { ZetkinCustomField, ZetkinPerson } from 'utils/types/zetkin';
 
 export interface ProfilesStoreSlice {
+  fieldsList: RemoteList<ZetkinCustomField>;
   personById: Record<number, RemoteItem<ZetkinPerson>>;
 }
 
 const initialState: ProfilesStoreSlice = {
+  fieldsList: remoteList(),
   personById: {},
 };
 
@@ -14,6 +21,13 @@ const profilesSlice = createSlice({
   initialState,
   name: 'profiles',
   reducers: {
+    fieldsLoad: (state) => {
+      state.fieldsList.isLoading = true;
+    },
+    fieldsLoaded: (state, action: PayloadAction<ZetkinCustomField[]>) => {
+      state.fieldsList = remoteList(action.payload);
+      state.fieldsList.loaded = new Date().toISOString();
+    },
     personLoad: (state, action: PayloadAction<number>) => {
       const id = action.payload;
       state.personById[id] = remoteItem(id, {
@@ -32,4 +46,5 @@ const profilesSlice = createSlice({
 });
 
 export default profilesSlice;
-export const { personLoad, personLoaded } = profilesSlice.actions;
+export const { fieldsLoad, fieldsLoaded, personLoad, personLoaded } =
+  profilesSlice.actions;
