+++ src/features/smartSearch/components/filters/PersonField/DisplayPersonField.tsx
@@ -10,7 +10,7 @@ import {
 import messageIds from 'features/smartSearch/l10n/messageIds';
 import UnderlinedMsg from '../../UnderlinedMsg';
 import UnderlinedText from '../../UnderlinedText';
-import useCustomFields from 'features/smartSearch/hooks/useCustomFields';
+import useCustomFields from 'features/profile/hooks/useCustomFields';
 import { useNumericRouteParams } from 'core/hooks';
 const localMessageIds = messageIds.filters.personField;
