+++ src/features/smartSearch/components/filters/PersonField/index.tsx
@@ -17,7 +17,7 @@ import {
 } from 'features/smartSearch/components/types';
 
 import messageIds from 'features/smartSearch/l10n/messageIds';
-import useCustomFields from 'features/smartSearch/hooks/useCustomFields';
+import useCustomFields from 'features/profile/hooks/useCustomFields';
 import { useNumericRouteParams } from 'core/hooks';
 const localMessageIds = messageIds.filters.personField;
