+++ src/features/views/components/ViewColumnDialog/PersonFieldConfig.tsx
@@ -21,7 +21,7 @@ import { Msg, useMessages } from 'core/i18n';
 
 import globalMessageIds from 'core/i18n/globalMessageIds';
 import messageIds from 'features/views/l10n/messageIds';
-import useCustomFields from 'features/smartSearch/hooks/useCustomFields';
+import useCustomFields from 'features/profile/hooks/useCustomFields';
 import { useNumericRouteParams } from 'core/hooks';
 
 interface PersonFieldConfigProps {
